patch: Move ca functions to utils/certificates
